Extract base class for the Filter presenter
This is a preliminary step to reuse it in the tree-based filter view
